+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 --------------------------------------------------------------------------------
+
+* support yelets imports, project functions are now available under `@import("project")`
+* support yelets if, for, while
+
+--------------------------------------------------------------------------------
 0.2.0, 26 September 2025
 
 * status, commit, push and update commands